move getLineUnderCursor to editor utils
